requirements.txt
pytest==7.1.2
requests==2.28.1
selenium==4.4.3
allure-pytest==2.10.0
allure-python-commons==2.10.0
pydantic==1.10.2
